[Chore] Reorganize engine.ts
